+++ packages/io-sign-issuer-sdk/package.json
@@ -40,7 +40,15 @@
     "whatwg-fetch": "^3.0.0"
   },
   "devDependencies": {
+    "@jest/globals": "^29.3.1",
+    "@pagopa/eslint-config": "^3.0.0",
+    "@rushstack/eslint-patch": "^1.2.0",
     "@types/url-parse": "1.4.4",
-    "typescript": "^4.0"
+    "eslint": "^8.28.0",
+    "jest": "^29.3.1",
+    "prettier": "2.7.1",
+    "ts-jest": "^29.0.3",
+    "tsup": "^6.4.0",
+    "typescript": "^4.8.4"
   }
 }
